feat: prevent slice construction on host
This is done because dereferencing a host generated slice on the guest will yield incorrect data (offset info isn't passed to the guest)
